requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# AI/ML dependencies
google-generativeai>=0.3.0
openai>=1.0.0

# Web framework
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0

# Utilities
python-dotenv>=1.0.0
pydantic-settings>=2.0.0
httpx>=0.25.0
aiofiles>=23.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.25.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
